ComToTcpServer: Send messages and blocks on COM
When sending large packets, the device FIFO may not be emptied fast enough
and data can end up getting lost. Patina can handle this to some degree,
but if the data lost is the checksum bytes, the stub does not realize the
packet was finished and will just wait indefinitely for more data.

This change forces the data to be sent in smaller blocks with delays in
between to allow the device to catch up. Ideally this is replaced by
proper flow control in the future.